Update index.html
initial setup of HTML for starting Responsive Web Lab. adding CSS Media Queries for max-width 1080 (1080x1920). at this time a Javascript file already existing in this repository for Hamburger functionality for Responsive Web, but we did'nt added that respective elements into HTML and CSS till this updation. we will add it soon. because we begin from one.